EW: Click count updates on avatar click action
